test: split test_sdlc_router_decision into per-theme modules (Refs #2879)
