+++ tests/unit/sdlc_router_decision/__init__.py
@@ -1,0 +1,1 @@
+"""Pure-function tests for agent.sdlc_router.decide_next_dispatch(), split by theme (#2879)."""
