--- tests/unit/test_sdlc_router_decision.py
@@ -1,2096 +1,0 @@
-"""Pure-function tests for agent.sdlc_router.decide_next_dispatch()."""
-
-from __future__ import annotations
-
-from agent.sdlc_router import (
-    DISPATCH_RULES,
-    MAX_CONCERN_RECRITIQUE_ROUNDS,
-    MAX_SAME_STAGE_DISPATCHES,
-    SKILL_DO_BUILD,
-    SKILL_DO_DOCS,
-    SKILL_DO_MERGE,
-    SKILL_DO_PATCH,
-    SKILL_DO_PLAN,
-    SKILL_DO_PLAN_CRITIQUE,
-    SKILL_DO_PR_REVIEW,
-    Blocked,
-    Dispatch,
-    _concern_revision_is_unjudged,
-    _critique_verdict_is_stale,
-    _review_verdict_is_stale,
-    _rule_patch_applied_after_review,
-    decide_next_dispatch,
-)
-
-
-def _states_all_pending() -> dict:
-    return {
-        "ISSUE": "pending",
-        "PLAN": "pending",
-        "CRITIQUE": "pending",
-        "BUILD": "pending",
-        "TEST": "pending",
-        "REVIEW": "pending",
-        "DOCS": "pending",
-        "MERGE": "pending",
-    }
-
-
-class TestDispatchRulesTable:
-    """Baseline: DISPATCH_RULES wiring is well-formed."""
-
-    def test_rules_have_unique_row_ids(self):
-        row_ids = [r.row_id for r in DISPATCH_RULES]
-        assert len(row_ids) == len(set(row_ids)), "duplicate row_id in DISPATCH_RULES"
-
-    def test_every_rule_has_a_docstring(self):
-        for rule in DISPATCH_RULES:
-            assert rule.state_predicate.__doc__, (
-                f"rule {rule.row_id} predicate missing __doc__ — parity test will fail"
-            )
-
-    def test_every_skill_is_known(self):
-        known = {
-            SKILL_DO_PLAN,
-            SKILL_DO_PLAN_CRITIQUE,
-            SKILL_DO_BUILD,
-            SKILL_DO_PATCH,
-            SKILL_DO_PR_REVIEW,
-            SKILL_DO_DOCS,
-            SKILL_DO_MERGE,
-        }
-        for rule in DISPATCH_RULES:
-            assert rule.skill in known, f"unknown skill {rule.skill} in row {rule.row_id}"
-
-
-class TestRow1NoPlan:
-    def test_empty_state_returns_do_plan(self):
-        result = decide_next_dispatch({}, {})
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN
-        assert result.row_id == "1"
-
-    def test_all_pending_returns_do_plan(self):
-        result = decide_next_dispatch(_states_all_pending(), {})
-        assert result.skill == SKILL_DO_PLAN
-
-
-class TestRow2PlanNotCritiqued:
-    def test_plan_completed_critique_pending_dispatches_critique(self):
-        states = _states_all_pending()
-        states["PLAN"] = "completed"
-        result = decide_next_dispatch(states, {})
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-        assert result.row_id == "2"
-
-    def test_plan_completed_critique_ready_dispatches_critique(self):
-        # When PLAN completes, the state machine auto-transitions CRITIQUE
-        # from "pending" to "ready". Row 2 must accept the "ready" status
-        # too (regression for the issue #1262 dispatch failure).
-        states = _states_all_pending()
-        states["PLAN"] = "completed"
-        states["CRITIQUE"] = "ready"
-        result = decide_next_dispatch(states, {})
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-        assert result.row_id == "2"
-
-
-class TestRow3CritiqueNeedsRevision:
-    def test_needs_revision_without_loop_dispatches_plan(self):
-        # G1 only trips when last skill was /do-plan-critique; without that,
-        # the dispatch table's Row 3 still routes back to /do-plan.
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "failed",
-            "_verdicts": {"CRITIQUE": {"verdict": "NEEDS REVISION"}},
-        }
-        result = decide_next_dispatch(states, {"latest_critique_verdict": "NEEDS REVISION"})
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN
-
-
-class TestRow4aReadyNoConcerns:
-    def test_ready_no_concerns_dispatches_build(self):
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "pending",
-        }
-        meta = {"latest_critique_verdict": "READY TO BUILD (no concerns)"}
-        result = decide_next_dispatch(states, meta)
-        assert result.skill == SKILL_DO_BUILD
-        assert result.row_id == "4a"
-
-
-class TestRow4bReadyWithConcernsNoRevision:
-    def test_concerns_without_revision_flag_returns_to_plan(self):
-        states = {"PLAN": "completed", "CRITIQUE": "completed"}
-        meta = {
-            "latest_critique_verdict": "READY TO BUILD (with concerns)",
-            "revision_applied": False,
-        }
-        result = decide_next_dispatch(states, meta)
-        assert result.skill == SKILL_DO_PLAN
-        assert result.row_id == "4b"
-
-    def test_releases_once_pr_exists_even_without_revision_flag(self):
-        """Row 4b must NOT re-dispatch /do-plan after the PR is open.
-
-        Symmetry gap: row 4c got pr_number/BUILD guards but row 4b did not, so
-        a with-concerns plan whose revision_applied flag never got set would
-        loop /do-plan forever once a PR existed, never advancing to review.
-        """
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "REVIEW": "ready",
-        }
-        meta = {
-            "latest_critique_verdict": "READY TO BUILD (with concerns)",
-            "revision_applied": False,
-            "pr_number": 466,
-        }
-        result = decide_next_dispatch(states, meta)
-        assert result.skill != SKILL_DO_PLAN
-        assert result.row_id != "4b"
-
-
-class TestRow4cReadyWithConcernsRevisionApplied:
-    def test_sticky_revision_flag_alone_no_longer_reaches_build(self):
-        """#2787: the sticky boolean is not evidence that THIS round was revised.
-
-        /do-plan sets `revision_applied: true` on every revision pass and it
-        never resets, so from round 2 onward this row fired unconditionally and
-        sent each new with-concerns verdict straight to /do-build unreviewed.
-        Row 4c is now the BOUND-EXHAUSTED build edge: it requires an
-        event-scoped revision AND a spent bound. With neither, routing falls to
-        row 4b for the revision pass.
-        """
-        states = {"PLAN": "completed", "CRITIQUE": "completed"}
-        meta = {
-            "latest_critique_verdict": "READY TO BUILD (with concerns)",
-            "revision_applied": True,
-        }
-        result = decide_next_dispatch(states, meta)
-        assert result.skill == SKILL_DO_PLAN
-        assert result.row_id == "4b"
-
-    def test_releases_to_review_once_pr_exists(self):
-        """Row 4c must NOT re-dispatch /do-build after the PR is open.
-
-        Regression: row 4c lacked the pr_number/BUILD guards that row 4a has,
-        so a with-concerns plan with revision_applied=True looped /do-build
-        forever and never advanced to review.
-        """
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "REVIEW": "ready",
-        }
-        meta = {
-            "latest_critique_verdict": "READY TO BUILD (with concerns)",
-            "revision_applied": True,
-            "pr_number": 466,
-        }
-        result = decide_next_dispatch(states, meta)
-        assert result.skill != SKILL_DO_BUILD
-        assert result.row_id != "4c"
-
-
-class TestD3FinishedPrNeverRoutesBackToBuild:
-    """D3: rows 4b/4c defer once a PR exists or BUILD completed."""
-
-    def test_4b_defers_when_pr_number_set(self):
-        states = {"PLAN": "completed", "CRITIQUE": "completed", "REVIEW": "pending"}
-        meta = {
-            "latest_critique_verdict": "READY TO BUILD (with concerns)",
-            "revision_applied": False,
-            "pr_number": 99,
-        }
-        result = decide_next_dispatch(states, meta)
-        # Must NOT route back to plan/build; a PR exists so downstream owns it.
-        assert result.skill != SKILL_DO_PLAN
-        assert result.skill != SKILL_DO_BUILD
-
-    def test_4c_defers_when_build_completed(self):
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-        }
-        meta = {
-            "latest_critique_verdict": "READY TO BUILD (with concerns)",
-            "revision_applied": True,
-        }
-        result = decide_next_dispatch(states, meta)
-        # 4c must not re-propose build once BUILD is completed. With no PR yet
-        # and no downstream rule matching, the router defers (Blocked) rather
-        # than routing back to /do-build.
-        assert getattr(result, "skill", None) != SKILL_DO_BUILD
-
-    def test_4c_finished_pr_routes_to_review(self):
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "REVIEW": "pending",
-        }
-        meta = {
-            "latest_critique_verdict": "READY TO BUILD (with concerns)",
-            "revision_applied": True,
-            "pr_number": 123,
-        }
-        result = decide_next_dispatch(states, meta)
-        assert result.skill == SKILL_DO_PR_REVIEW
-
-
-class TestRow6TestsFailing:
-    def test_test_failed_dispatches_patch(self):
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "failed",
-        }
-        result = decide_next_dispatch(states, {})
-        assert result.skill == SKILL_DO_PATCH
-        assert result.row_id == "6"
-
-
-class TestRow7PrExistsNoReview:
-    def test_pr_no_review_dispatches_review(self):
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "REVIEW": "pending",
-        }
-        meta = {"pr_number": 1234}
-        result = decide_next_dispatch(states, meta)
-        assert result.skill == SKILL_DO_PR_REVIEW
-        assert result.row_id == "7"
-
-
-class TestRow8ReviewHasFindings:
-    def test_changes_requested_dispatches_patch(self):
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "REVIEW": "failed",
-        }
-        meta = {"pr_number": 99, "latest_review_verdict": "CHANGES REQUESTED"}
-        result = decide_next_dispatch(states, meta)
-        assert result.skill == SKILL_DO_PATCH
-        assert result.row_id == "8"
-
-
-class TestRow8bPatchAppliedAfterReview:
-    def test_patch_complete_after_review_triggers_rereview(self):
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "REVIEW": "failed",
-            "PATCH": "completed",
-        }
-        meta = {"pr_number": 99, "last_dispatched_skill": SKILL_DO_PATCH}
-        result = decide_next_dispatch(states, meta)
-        assert result.skill == SKILL_DO_PR_REVIEW
-        assert result.row_id == "8b"
-
-
-class TestRow9ReviewApprovedDocsNotDone:
-    def test_review_completed_docs_pending_dispatches_docs(self):
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "REVIEW": "completed",
-            "DOCS": "pending",
-        }
-        # #1932 gap (c): row 9 now requires a recorded APPROVED verdict, not
-        # just REVIEW==completed — record one so this exercises row 9's real
-        # (fixed) contract instead of the old verdict-blind misroute.
-        meta = {"pr_number": 7, "latest_review_verdict": "APPROVED"}
-        result = decide_next_dispatch(states, meta)
-        assert result.skill == SKILL_DO_DOCS
-        assert result.row_id == "9"
-
-
-class TestRow10ReadyToMerge:
-    def test_all_completed_dispatches_merge(self):
-        states = {
-            "ISSUE": "completed",
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "REVIEW": "completed",
-            "DOCS": "completed",
-        }
-        # WS3a (#2062): row 10 requires a recorded APPROVED verdict, mirroring
-        # row 9 -- REVIEW==completed alone is no longer merge-ready.
-        meta = {"pr_number": 42, "latest_review_verdict": "APPROVED"}
-        result = decide_next_dispatch(states, meta)
-        assert result.skill == SKILL_DO_MERGE
-        assert result.row_id == "10"
-
-    def test_all_completed_without_verdict_routes_to_review_not_merge(self):
-        """WS3a/b (#2062): the same all-completed state with NO recorded
-        verdict must re-review (row 8e), never merge."""
-        states = {
-            "ISSUE": "completed",
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "REVIEW": "completed",
-            "DOCS": "completed",
-        }
-        meta = {"pr_number": 42}
-        result = decide_next_dispatch(states, meta)
-        assert result.skill == SKILL_DO_PR_REVIEW
-        assert result.row_id == "8e"
-
-
-class TestStageStatesUnavailableNoMergeDispatch:
-    """Row 10b deleted (#2003): stage_states-unavailable + open PR must never
-    fall back to dispatching /do-merge. Enforcement of the merge predicate
-    lives in the merge-guard hook (tools.merge_predicate); the router only
-    schedules merge via row 10 (all stages confirmed completed) or G6.
-    """
-
-    def test_row_10b_removed_from_dispatch_rules(self):
-        assert "10b" not in {r.row_id for r in DISPATCH_RULES}
-
-    def test_empty_states_pr_open_routes_to_review_not_merge(self):
-        # Without stage_states we cannot confirm docs/review are done, so the
-        # safest dispatch is /do-pr-review — never merge.
-        result = decide_next_dispatch({}, {"pr_number": 1234})
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PR_REVIEW
-        assert result.row_id == "7"
-
-    def test_empty_states_past_docs_never_dispatches_merge(self):
-        # Pre-#2003 this state could reach row 10b's merge fallback. Now the
-        # router must either dispatch a non-merge skill or report Blocked —
-        # stage_states-unavailable is never sufficient evidence to merge.
-        result = decide_next_dispatch(
-            {},
-            {
-                "pr_number": 1234,
-                "latest_review_verdict": "APPROVED",
-                "last_dispatched_skill": SKILL_DO_DOCS,
-            },
-        )
-        if isinstance(result, Dispatch):
-            assert result.skill != SKILL_DO_MERGE
-        else:
-            assert isinstance(result, Blocked)
-
-
-class TestNoMatchingRule:
-    def test_impossible_state_returns_blocked(self):
-        # Craft a state where no rule matches: PLAN completed but CRITIQUE also
-        # completed AND no verdict AND BUILD completed AND TEST completed AND
-        # REVIEW completed AND DOCS completed but NO pr_number.
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "REVIEW": "completed",
-            "DOCS": "completed",
-        }
-        result = decide_next_dispatch(states, {})  # no pr_number => Row 10 fails
-        assert isinstance(result, Blocked)
-
-
-# ---------------------------------------------------------------------------
-# #1638 — verdict normalization: underscore forms match space-canonical rules
-# ---------------------------------------------------------------------------
-
-
-class TestVerdictNormalizationUnderscore:
-    """Underscore-form verdicts must route identically to space-form (#1638)."""
-
-    def _review_states_with_findings(self) -> dict:
-        return {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "REVIEW": "failed",
-        }
-
-    def test_review_changes_requested_underscore_dispatches_patch(self):
-        states = self._review_states_with_findings()
-        meta = {"pr_number": 99, "latest_review_verdict": "CHANGES_REQUESTED"}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PATCH
-        assert result.row_id == "8"
-
-    def test_review_changes_requested_space_dispatches_patch(self):
-        states = self._review_states_with_findings()
-        meta = {"pr_number": 99, "latest_review_verdict": "CHANGES REQUESTED"}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PATCH
-        assert result.row_id == "8"
-
-
-# ---------------------------------------------------------------------------
-# #1640 — plan-existence evidence gate: PLAN="ready" needs a real plan doc
-# ---------------------------------------------------------------------------
-
-
-class TestPlanExistenceGate:
-    """PLAN='ready' without a plan doc must route to /do-plan, not /do-plan-critique."""
-
-    def test_fresh_session_plan_ready_no_doc_dispatches_do_plan(self):
-        """Bootstrap: PLAN='ready', no plan doc → dispatch /do-plan (Row 1)."""
-        states = {"PLAN": "ready", "CRITIQUE": None}
-        meta = {"plan_exists": False, "issue_number": 1234}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN
-
-    def test_plan_ready_plan_exists_dispatches_critique(self):
-        """PLAN='ready' WITH a plan doc → dispatch /do-plan-critique (Row 2)."""
-        states = {"PLAN": "ready", "CRITIQUE": None}
-        meta = {"plan_exists": True, "issue_number": 1234}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-
-    def test_plan_ready_no_issue_number_does_not_reroute_to_plan(self):
-        """PLAN='ready', plan_exists=False, issue_number=None → falls through to Row 2.
-
-        Without an issue_number we cannot verify the plan file, so we treat
-        the 'ready' state as an implicit plan exists and let Row 2 run critique.
-        The guard only fires when issue_number is available for a lookup.
-        """
-        states = {"PLAN": "ready", "CRITIQUE": None}
-        meta = {"plan_exists": False, "issue_number": None}
-        result = decide_next_dispatch(states, meta)
-        # Without issue_number, _rule_no_plan does NOT fire for PLAN="ready";
-        # the state should fall through to Row 2 (critique) or Blocked.
-        assert isinstance(result, (Dispatch, Blocked))
-        if isinstance(result, Dispatch):
-            assert result.skill != SKILL_DO_PLAN
-
-    def test_plan_completed_always_dispatches_critique_regardless_of_plan_exists(self):
-        """PLAN='completed' → always critique, regardless of plan_exists (#1275 intact)."""
-        states = {"PLAN": "completed", "CRITIQUE": None}
-        meta = {"plan_exists": False, "issue_number": 9999}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-
-
-# ---------------------------------------------------------------------------
-# #1641 — patch-supersedes-stale-verdict timestamp early-exit
-# ---------------------------------------------------------------------------
-
-
-def _iso(ts: str) -> str:
-    """Return an ISO-8601 timestamp string (thin wrapper for readability)."""
-    return ts
-
-
-class TestReviewVerdictStaleness:
-    """Stale REVIEW verdict (older than latest /do-patch dispatch) must be superseded."""
-
-    def _base_states(self, verdict_at: str, patch_at: str) -> dict:
-        return {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "REVIEW": "failed",
-            "PATCH": "completed",
-            "_verdicts": {
-                "REVIEW": {
-                    "verdict": "CHANGES REQUESTED",
-                    "recorded_at": verdict_at,
-                }
-            },
-            "_sdlc_dispatches": [
-                {"skill": "/do-patch", "at": patch_at},
-            ],
-        }
-
-    def test_stale_review_verdict_after_patch_dispatches_review(self):
-        """verdict T0 < patch T1 → stale → /do-pr-review (row 8b)."""
-        states = self._base_states(
-            verdict_at="2026-01-01T10:00:00",
-            patch_at="2026-01-01T11:00:00",
-        )
-        meta = {"pr_number": 99, "last_dispatched_skill": SKILL_DO_PATCH}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PR_REVIEW
-
-    def test_fresh_review_verdict_after_patch_dispatches_patch(self):
-        """verdict T2 > patch T1 → fresh → /do-patch (row 8)."""
-        states = self._base_states(
-            verdict_at="2026-01-01T12:00:00",
-            patch_at="2026-01-01T11:00:00",
-        )
-        meta = {"pr_number": 99, "latest_review_verdict": "CHANGES REQUESTED"}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PATCH
-
-    def test_review_verdict_stale_missing_recorded_at_not_suppressed(self):
-        """Missing recorded_at → not stale → row 8 fires normally."""
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "REVIEW": "failed",
-            "PATCH": "completed",
-            "_verdicts": {
-                "REVIEW": {
-                    "verdict": "CHANGES REQUESTED",
-                    # no recorded_at
-                }
-            },
-            "_sdlc_dispatches": [{"skill": "/do-patch", "at": "2026-01-01T11:00:00"}],
-        }
-        meta = {"pr_number": 99, "latest_review_verdict": "CHANGES REQUESTED"}
-        result = decide_next_dispatch(states, meta)
-        # Not stale (missing recorded_at) → row 8 fires → /do-patch
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PATCH
-
-    def test_review_verdict_stale_no_prior_patch_not_suppressed(self):
-        """No /do-patch in dispatch history → not stale → row 8 fires."""
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "REVIEW": "failed",
-            "_verdicts": {
-                "REVIEW": {
-                    "verdict": "CHANGES REQUESTED",
-                    "recorded_at": "2026-01-01T10:00:00",
-                }
-            },
-            "_sdlc_dispatches": [],  # no /do-patch entries
-        }
-        meta = {"pr_number": 99, "latest_review_verdict": "CHANGES REQUESTED"}
-        result = decide_next_dispatch(states, meta)
-        # No prior /do-patch → not stale → row 8 fires → /do-patch
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PATCH
-
-    def test_review_verdict_stale_equal_timestamps_fresh(self):
-        """Equal timestamps → not stale (strict <) → row 8 fires."""
-        ts = "2026-01-01T10:00:00"
-        states = self._base_states(verdict_at=ts, patch_at=ts)
-        meta = {"pr_number": 99, "latest_review_verdict": "CHANGES REQUESTED"}
-        result = decide_next_dispatch(states, meta)
-        # Equal timestamps → not stale → row 8 → /do-patch
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PATCH
-
-    def test_review_verdict_stale_non_iso_timestamp_not_suppressed(self):
-        """Malformed recorded_at → parse failure → not stale → row 8 fires."""
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "REVIEW": "failed",
-            "_verdicts": {
-                "REVIEW": {
-                    "verdict": "CHANGES REQUESTED",
-                    "recorded_at": "not-a-date",
-                }
-            },
-            "_sdlc_dispatches": [{"skill": "/do-patch", "at": "2026-01-01T11:00:00"}],
-        }
-        meta = {"pr_number": 99, "latest_review_verdict": "CHANGES REQUESTED"}
-        result = decide_next_dispatch(states, meta)
-        # Parse failure → not stale → row 8 → /do-patch
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PATCH
-
-    def test_fresh_review_verdict_after_patch_still_dispatches_patch(self):
-        """patch T1 < verdict T2 → fresh verdict → row 8 wins → /do-patch."""
-        states = self._base_states(
-            verdict_at="2026-01-01T12:00:00",
-            patch_at="2026-01-01T11:00:00",
-        )
-        # Override latest_review_verdict directly so meta takes precedence.
-        meta = {"pr_number": 99, "latest_review_verdict": "CHANGES REQUESTED"}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PATCH
-
-
-class TestCritiqueVerdictStaleness:
-    """Stale CRITIQUE verdict (older than latest /do-plan dispatch) must re-critique (#1639).
-
-    Mirrors TestReviewVerdictStaleness for the CRITIQUE path. A NEEDS REVISION
-    verdict recorded before the plan was revised (a later /do-plan dispatch)
-    routes to /do-plan-critique (row 2b) instead of dead-ending on /do-plan
-    (row 3).
-    """
-
-    def _base_states(self, verdict_at: str, plan_at: str) -> dict:
-        return {
-            "PLAN": "completed",
-            "CRITIQUE": "in_progress",
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "NEEDS REVISION",
-                    "recorded_at": verdict_at,
-                }
-            },
-            "_sdlc_dispatches": [
-                {"skill": "/do-plan", "at": plan_at},
-            ],
-        }
-
-    def test_stale_critique_verdict_after_plan_dispatches_recritique(self):
-        """verdict T0 < plan T1 → stale → /do-plan-critique (row 2b)."""
-        states = self._base_states(
-            verdict_at="2026-01-01T10:00:00",
-            plan_at="2026-01-01T11:00:00",
-        )
-        meta = {"last_dispatched_skill": SKILL_DO_PLAN}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-
-    def test_fresh_critique_verdict_after_plan_dispatches_plan(self):
-        """verdict T2 > plan T1 → fresh → /do-plan (row 3), no over-suppression."""
-        states = self._base_states(
-            verdict_at="2026-01-01T12:00:00",
-            plan_at="2026-01-01T11:00:00",
-        )
-        meta = {"last_dispatched_skill": SKILL_DO_PLAN}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN
-
-    def test_critique_verdict_stale_missing_recorded_at_not_suppressed(self):
-        """Missing recorded_at → not stale → row 3 fires (/do-plan)."""
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "in_progress",
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "NEEDS REVISION",
-                    # no recorded_at
-                }
-            },
-            "_sdlc_dispatches": [{"skill": "/do-plan", "at": "2026-01-01T11:00:00"}],
-        }
-        meta = {"last_dispatched_skill": SKILL_DO_PLAN}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN
-
-    def test_critique_verdict_stale_no_prior_plan_not_suppressed(self):
-        """No /do-plan in dispatch history → not stale → row 3 fires."""
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "in_progress",
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "NEEDS REVISION",
-                    "recorded_at": "2026-01-01T10:00:00",
-                }
-            },
-            "_sdlc_dispatches": [],  # no /do-plan entries
-        }
-        meta = {"last_dispatched_skill": SKILL_DO_PLAN}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN
-
-    def test_critique_verdict_equal_timestamps_fresh(self):
-        """Equal timestamps → not stale (strict <) → row 3 fires (/do-plan)."""
-        ts = "2026-01-01T10:00:00"
-        states = self._base_states(verdict_at=ts, plan_at=ts)
-        meta = {"last_dispatched_skill": SKILL_DO_PLAN}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN
-
-    def test_critique_verdict_stale_non_iso_timestamp_not_suppressed(self):
-        """Malformed recorded_at → parse failure → not stale → row 3 fires."""
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "in_progress",
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "NEEDS REVISION",
-                    "recorded_at": "not-a-date",
-                }
-            },
-            "_sdlc_dispatches": [{"skill": "/do-plan", "at": "2026-01-01T11:00:00"}],
-        }
-        meta = {"last_dispatched_skill": SKILL_DO_PLAN}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN
-
-    def test_combined_sourcing_dead_end_routes_recritique(self):
-        """#1639 dead-end: BOTH _verdicts.CRITIQUE and meta.latest_critique_verdict set.
-
-        Staleness sources recorded_at from _verdicts only; row-3 text sources from
-        meta. Populate both so the test cannot pass for the wrong reason (Concern 2).
-        """
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "in_progress",
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "NEEDS REVISION",
-                    "recorded_at": "2026-01-01T10:00:00",  # T0
-                }
-            },
-            "_sdlc_dispatches": [
-                {"skill": "/do-plan", "at": "2026-01-01T11:00:00"},  # T1 > T0
-            ],
-        }
-        meta = {
-            "latest_critique_verdict": "NEEDS REVISION",
-            "last_dispatched_skill": SKILL_DO_PLAN,
-            "revision_applied": True,
-        }
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-
-
-class TestConvergenceLatchRevisionAppliedAt:
-    """Event-scoped convergence latch (#1760).
-
-    A bare ``revision_applied`` boolean can't distinguish "this is the
-    settle-and-build dispatch" from "this is some later unrelated /do-plan
-    dispatch" -- /do-plan sets it true on every revision pass. Pairing it
-    with an event-scoped ``meta["revision_applied_at"]`` timestamp lets
-    ``_critique_verdict_is_stale`` suppress staleness only when the latest
-    ``/do-plan`` dispatch is NOT LATER than ``revision_applied_at``.
-    """
-
-    def test_1760_loop_replay_settles_to_build(self):
-        """The exact #1760 loop replay: notes-only revision converges to
-        /do-build instead of re-dispatching /do-plan-critique forever.
-
-        READY TO BUILD verdict recorded at T0. A notes-only revision runs
-        /do-plan (dispatch at T0.5, postdating the verdict) and sets
-        revision_applied_at to T1 (postdating the dispatch). The latch must
-        suppress staleness so row 2b steps aside and row 4a routes to
-        /do-build.
-        """
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "in_progress",
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "READY TO BUILD",
-                    "recorded_at": "2026-01-01T10:00:00",  # T0
-                }
-            },
-            "_sdlc_dispatches": [
-                {"skill": "/do-plan", "at": "2026-01-01T10:30:00"},  # T0.5 > T0
-            ],
-        }
-        meta = {
-            "last_dispatched_skill": SKILL_DO_PLAN,
-            "latest_critique_verdict": "READY TO BUILD",
-            "revision_applied": True,
-            "revision_applied_at": "2026-01-01T11:00:00",  # T1 > T0.5 (dispatch)
-        }
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_BUILD
-        assert result.skill != SKILL_DO_PLAN_CRITIQUE
-
-    def test_predicate_returns_false_on_settle_and_build_case(self):
-        """Isolated-predicate test: calling _critique_verdict_is_stale directly
-        (not via the row-2b wrapper) proves the PREDICATE changed, not just a
-        caller.
-        """
-        states = {
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "READY TO BUILD",
-                    "recorded_at": "2026-01-01T10:00:00",  # T0
-                }
-            },
-            "_sdlc_dispatches": [
-                {"skill": "/do-plan", "at": "2026-01-01T10:30:00"},  # T0.5 > T0
-            ],
-        }
-        meta = {"revision_applied_at": "2026-01-01T11:00:00"}  # T1 > T0.5
-        assert _critique_verdict_is_stale(states, meta) is False
-
-    def test_later_unrelated_dispatch_re_stales_normally(self):
-        """A /do-plan dispatch postdating revision_applied_at must re-stale
-        normally -- a genuinely-stale verdict must not be routed to BUILD
-        just because revision_applied happens to still read true.
-        """
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "in_progress",
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "READY TO BUILD",
-                    "recorded_at": "2026-01-01T10:00:00",  # T0
-                }
-            },
-            "_sdlc_dispatches": [
-                {"skill": "/do-plan", "at": "2026-01-02T00:00:00"},  # T2 > T1
-            ],
-        }
-        meta = {
-            "last_dispatched_skill": SKILL_DO_PLAN,
-            "latest_critique_verdict": "READY TO BUILD",
-            "revision_applied": True,
-            "revision_applied_at": "2026-01-01T11:00:00",  # T1 < T2 (dispatch)
-        }
-        assert _critique_verdict_is_stale(states, meta) is True
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-
-    def test_predicate_malformed_recorded_at_still_false(self):
-        """Existing fail-safe preserved: malformed recorded_at → False, even
-        with a well-formed revision_applied_at present in meta.
-        """
-        states = {
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "READY TO BUILD",
-                    "recorded_at": "not-a-date",
-                }
-            },
-            "_sdlc_dispatches": [{"skill": "/do-plan", "at": "2026-01-01T11:00:00"}],
-        }
-        meta = {"revision_applied_at": "2026-01-01T11:30:00"}
-        assert _critique_verdict_is_stale(states, meta) is False
-
-    def test_predicate_malformed_revision_applied_at_falls_back_to_timestamp_only(self):
-        """Malformed/absent revision_applied_at leaves the latch inert -- the
-        predicate falls back to the pre-#1760 timestamp-only staleness check
-        (never wrongly reports "not stale" when the latch signal is missing).
-        """
-        states = {
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "READY TO BUILD",
-                    "recorded_at": "2026-01-01T10:00:00",  # T0
-                }
-            },
-            "_sdlc_dispatches": [
-                {"skill": "/do-plan", "at": "2026-01-01T11:00:00"},  # T1 > T0
-            ],
-        }
-        # Malformed revision_applied_at -> latch inert -> normal staleness (True).
-        assert _critique_verdict_is_stale(states, {"revision_applied_at": "not-a-date"}) is True
-        # Absent revision_applied_at -> latch inert -> normal staleness (True).
-        assert _critique_verdict_is_stale(states, {}) is True
-        assert _critique_verdict_is_stale(states, None) is True
-
-
-class TestCritiqueInProgressNoVerdictDeadEnd:
-    """CRITIQUE in_progress with NO recorded verdict must re-dispatch /do-plan-critique (#1668).
-
-    Reconstructed from the real issue-1643 run (PR #1664). The
-    ``/do-plan-critique`` skill ran but never persisted a verdict to
-    ``_verdicts.CRITIQUE`` — the war-room cycles happened in-skill and the
-    verdict-record step never fired, leaving CRITIQUE stuck at ``in_progress``.
-    With no verdict text and no ``recorded_at``:
-
-    - row 2 (``_rule_plan_not_critiqued``) excludes ``in_progress`` → no match
-    - row 2b (``_rule_critique_verdict_stale``) needs ``recorded_at`` → no match
-    - rows 3/4a/4b/4c need a recorded verdict → no match
-    - G1 needs NEEDS REVISION / MAJOR REWORK verdict text → no match
-
-    Result before the fix: ``Blocked('no matching dispatch rule')`` — the exact
-    dead-end the supervisor navigated manually by re-running /do-plan-critique.
-    This is a distinct variant from the row-2b NEEDS-REVISION staleness case.
-    """
-
-    def _dead_end_states(self) -> dict:
-        """The exact reconstructed issue-1643 state at the dead-end moment."""
-        return {
-            "ISSUE": "ready",
-            "PLAN": "completed",
-            "CRITIQUE": "in_progress",
-            "BUILD": "pending",
-            "TEST": "pending",
-            "PATCH": "pending",
-            "REVIEW": "pending",
-            "DOCS": "pending",
-            "MERGE": "pending",
-            "_verdicts": {},  # critique ran but never recorded a verdict
-            "_sdlc_dispatches": [
-                {"skill": "/do-plan", "at": "2026-06-12T19:25:35.011316+00:00"},
-                {"skill": "/do-plan-critique", "at": "2026-06-12T19:32:07.722784+00:00"},
-            ],
-        }
-
-    def test_in_progress_no_verdict_no_pr_redispatches_critique(self):
-        """CRITIQUE in_progress + empty verdict + no PR → re-dispatch /do-plan-critique."""
-        states = self._dead_end_states()
-        meta = {
-            "issue_number": 1643,
-            "plan_exists": True,
-            "revision_applied": False,
-            "latest_critique_verdict": None,
-            "pr_number": None,
-            "last_dispatched_skill": SKILL_DO_PLAN_CRITIQUE,
-        }
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch), f"expected Dispatch, got {result!r}"
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-
-    def test_in_progress_no_verdict_with_pr_defers_to_pr_stage(self):
-        """Once a PR exists, the stalled-critique fast-path must NOT fire.
-
-        The real 1643 run progressed to a PR via PR-stage rows while CRITIQUE
-        stayed in_progress. A PR open means plan/critique are behind us — the
-        re-dispatch must defer to G3/PR-stage routing, never bounce back to
-        /do-plan-critique.
-        """
-        states = self._dead_end_states()
-        states["REVIEW"] = "pending"
-        meta = {
-            "issue_number": 1643,
-            "plan_exists": True,
-            "pr_number": 1664,
-            "last_dispatched_skill": SKILL_DO_PLAN_CRITIQUE,
-        }
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill != SKILL_DO_PLAN_CRITIQUE
-
-    def test_in_progress_with_recorded_verdict_unaffected(self):
-        """A recorded verdict must still route via rows 2b/3/4a — fast-path is verdict-gated."""
-        states = self._dead_end_states()
-        states["_verdicts"] = {
-            "CRITIQUE": {"verdict": "READY TO BUILD", "recorded_at": "2026-06-12T19:40:00+00:00"}
-        }
-        meta = {
-            "issue_number": 1643,
-            "plan_exists": True,
-            "latest_critique_verdict": "READY TO BUILD",
-            "pr_number": None,
-            "last_dispatched_skill": SKILL_DO_PLAN_CRITIQUE,
-        }
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        # READY TO BUILD, no concerns → row 4a → /do-build (NOT a re-critique)
-        assert result.skill == SKILL_DO_BUILD
-
-
-class TestCritiqueStaleG5LoopBound:
-    """Row 2b re-critique is bounded by G5 on an unchanged plan hash (Concern 1)."""
-
-    def _stale_states(self, artifact_hash: str | None = None) -> dict:
-        verdict: dict = {
-            "verdict": "NEEDS REVISION",
-            "recorded_at": "2026-01-01T10:00:00",  # T0
-        }
-        if artifact_hash is not None:
-            verdict["artifact_hash"] = artifact_hash
-        return {
-            "PLAN": "completed",
-            "CRITIQUE": "in_progress",
-            "_verdicts": {"CRITIQUE": verdict},
-            "_sdlc_dispatches": [
-                {"skill": "/do-plan", "at": "2026-01-01T11:00:00"},  # T1 > T0
-            ],
-        }
-
-    def test_stale_unchanged_hash_g5_short_circuits(self):
-        """Stale verdict + cached hash == current hash → G5 returns cached dispatch.
-
-        G5 fires before row 2b and, on an unchanged plan hash, routes the cached
-        NEEDS REVISION verdict to /do-plan (its downstream dispatch) — NOT a
-        re-critique loop.
-        """
-        states = self._stale_states(artifact_hash="sha256:abc")
-        meta = {"last_dispatched_skill": SKILL_DO_PLAN}
-        result = decide_next_dispatch(states, meta, context={"current_plan_hash": "sha256:abc"})
-        assert isinstance(result, Dispatch)
-        assert result.row_id == "G5"
-        assert result.skill == SKILL_DO_PLAN
-
-    def test_stale_changed_hash_dispatches_recritique(self):
-        """Stale verdict + cached hash != current hash → row 2b → /do-plan-critique."""
-        states = self._stale_states(artifact_hash="sha256:old")
-        meta = {"last_dispatched_skill": SKILL_DO_PLAN}
-        result = decide_next_dispatch(states, meta, context={"current_plan_hash": "sha256:new"})
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-
-
-class TestG1VsRow2bMutualExclusivity:
-    """G1 (last=critique) and row 2b (last=/do-plan, stale) never fire each other's skill."""
-
-    def _states(self) -> dict:
-        return {
-            "PLAN": "completed",
-            "CRITIQUE": "in_progress",
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "NEEDS REVISION",
-                    "recorded_at": "2026-01-01T10:00:00",
-                }
-            },
-            "_sdlc_dispatches": [
-                {"skill": "/do-plan", "at": "2026-01-01T11:00:00"},
-            ],
-        }
-
-    def test_g1_fires_when_last_dispatch_was_critique(self):
-        """last = /do-plan-critique → G1 → /do-plan (not /do-plan-critique)."""
-        states = self._states()
-        meta = {"last_dispatched_skill": SKILL_DO_PLAN_CRITIQUE}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN
-
-    def test_row2b_fires_when_last_dispatch_was_plan(self):
-        """last = /do-plan + stale verdict → row 2b → /do-plan-critique (not /do-plan)."""
-        states = self._states()
-        meta = {"last_dispatched_skill": SKILL_DO_PLAN}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-
-
-class TestReviewMarkerDesync:
-    """#1642: an APPROVED REVIEW verdict with a non-completed REVIEW marker must NOT advance docs.
-
-    Row 9 (_rule_review_approved_docs_not_done) requires REVIEW == completed in
-    stage_states. A desynced {verdict=APPROVED, marker!=completed} state is
-    router-observable: the router does NOT fire /do-docs, documenting that the
-    skill-layer completion-marker write is what advances REVIEW.
-    """
-
-    def test_approved_verdict_but_review_marker_not_completed_no_docs(self):
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "REVIEW": "in_progress",  # marker NOT completed despite APPROVED verdict
-            "DOCS": "pending",
-            "_verdicts": {
-                "REVIEW": {
-                    "verdict": "APPROVED",
-                    "recorded_at": "2026-01-01T10:00:00",
-                }
-            },
-        }
-        meta = {"pr_number": 99, "latest_review_verdict": "APPROVED"}
-        result = decide_next_dispatch(states, meta)
-        # Must NOT route to /do-docs (row 9 requires REVIEW == completed).
-        if isinstance(result, Dispatch):
-            assert result.skill != SKILL_DO_DOCS
-
-
-class TestReviewInProgressNoVerdictDeadEnd:
-    """REVIEW in_progress with NO recorded verdict must re-dispatch /do-pr-review (#1687).
-
-    Gap A (REVIEW path analogue of CRITIQUE row 2c / #1668): /do-pr-review ran
-    (REVIEW marker == "in_progress") but never persisted a verdict, so
-    _verdicts.REVIEW is empty and latest_review_verdict is None. With rows 7/8/8b
-    all missing (no verdict, no completed PATCH), the router fell through to
-    Blocked('no matching dispatch rule'). Row 8c recovers by re-dispatching
-    /do-pr-review.
-
-    Gap demonstration (concern #1): the UNFIXED router state (REVIEW in_progress,
-    empty verdict, no completed PATCH, PR present) is confirmed below via the
-    no-8c baseline fixture. The post-fix assertion documents the closed gap.
-    """
-
-    def _dead_end_states(self) -> dict:
-        """Reconstructed dead-end state: REVIEW in_progress, no verdict recorded."""
-        return {
-            "ISSUE": "ready",
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "TEST": "completed",
-            "PATCH": "pending",
-            "REVIEW": "in_progress",  # marker stuck — review ran but no verdict
-            "DOCS": "pending",
-            "MERGE": "pending",
-            "_verdicts": {},  # review ran but never recorded a verdict
-            "_sdlc_dispatches": [
-                {"skill": "/do-build", "at": "2026-06-14T10:00:00+00:00"},
-                {"skill": "/do-pr-review", "at": "2026-06-14T10:30:00+00:00"},
-            ],
-        }
-
-    def test_in_progress_no_verdict_pr_present_redispatches_review(self):
-        """REVIEW in_progress + empty verdict + PR present → re-dispatch /do-pr-review (row 8c).
-
-        This is the fixed state: what was previously Blocked('no matching dispatch rule')
-        (Gap A) now resolves to Dispatch(/do-pr-review) via new row 8c.
-        """
-        states = self._dead_end_states()
-        meta = {
-            "issue_number": 1687,
-            "pr_number": 9999,
-            "latest_review_verdict": None,
-            "last_dispatched_skill": SKILL_DO_PR_REVIEW,
-        }
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch), f"expected Dispatch, got {result!r}"
-        assert result.skill == SKILL_DO_PR_REVIEW
-        assert result.row_id == "8c"
-
-    def test_in_progress_no_verdict_no_pr_does_not_fire_8c(self):
-        """No PR → row 8c must NOT fire (REVIEW only exists post-PR).
-
-        Defensive safety gate: a REVIEW in_progress state with no PR is
-        structurally impossible in a well-ordered pipeline, but the predicate
-        guards against it explicitly.
-        """
-        states = self._dead_end_states()
-        states["REVIEW"] = "in_progress"
-        meta = {
-            "issue_number": 1687,
-            "pr_number": None,
-            "latest_review_verdict": None,
-            "last_dispatched_skill": SKILL_DO_PR_REVIEW,
-        }
-        result = decide_next_dispatch(states, meta)
-        # Row 8c requires pr_number; without it, must NOT fire 8c.
-        if isinstance(result, Dispatch):
-            assert result.row_id != "8c", "8c fired without a PR — predicate gate broken"
-
-    def test_whitespace_only_verdict_treated_as_empty_fires_8c(self):
-        """A whitespace-only verdict must be treated as empty so row 8c fires.
-
-        Mirrors the .strip() correctness requirement from critique concern #2:
-        a verdict of " " or "\n" must NOT read as "recorded" — it must fire 8c.
-        """
-        states = self._dead_end_states()
-        states["REVIEW"] = "in_progress"
-        states["_verdicts"] = {"REVIEW": {"verdict": "   ", "recorded_at": None}}
-        meta = {
-            "issue_number": 1687,
-            "pr_number": 9999,
-            "latest_review_verdict": "   ",  # whitespace only
-            "last_dispatched_skill": SKILL_DO_PR_REVIEW,
-        }
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch), f"expected Dispatch, got {result!r}"
-        assert result.skill == SKILL_DO_PR_REVIEW
-        assert result.row_id == "8c", f"expected row 8c, got {result.row_id!r}"
-
-    def test_patch_applied_after_review_defers_to_8b_not_8c(self):
-        """Row 8b (patch applied) owns states 8c must step aside from.
-
-        8c steps aside for 8b when PATCH==completed AND last_dispatched_skill==/do-patch
-        (the three-condition 8b predicate). Using _rule_patch_applied_after_review
-        as the gate (not a bare PATCH==completed check) is load-bearing.
-        """
-        states = self._dead_end_states()
-        states["REVIEW"] = "in_progress"
-        states["PATCH"] = "completed"
-        meta = {
-            "issue_number": 1687,
-            "pr_number": 9999,
-            "latest_review_verdict": None,
-            "last_dispatched_skill": SKILL_DO_PATCH,  # row 8b\'s third condition
-        }
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch), f"expected Dispatch, got {result!r}"
-        assert result.skill == SKILL_DO_PR_REVIEW
-        # Row 8b owns this state (patch applied after review), not 8c.
-        assert result.row_id == "8b", f"expected row 8b, got {result.row_id!r}"
-
-    def test_patch_completed_wrong_last_skill_8c_owns_state(self):
-        """PATCH completed but last_dispatched_skill != /do-patch → 8b does not own it, 8c must.
-
-        This is the "Blocked leak" prevention case: if we used a bare
-        PATCH==completed check in 8c, this state would fall through both 8b and 8c
-        to Blocked. Gating on _rule_patch_applied_after_review() exactly means 8c
-        correctly owns this case and re-dispatches /do-pr-review.
-        """
-        states = self._dead_end_states()
-        states["REVIEW"] = "in_progress"
-        states["PATCH"] = "completed"
-        meta = {
-            "issue_number": 1687,
-            "pr_number": 9999,
-            "latest_review_verdict": None,
-            "last_dispatched_skill": SKILL_DO_BUILD,  # NOT /do-patch — 8b skips it
-        }
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch), f"expected Dispatch, got {result!r}"
-        assert result.skill == SKILL_DO_PR_REVIEW
-        # 8b did not fire (wrong last_dispatched_skill), so 8c must own this state.
-        assert result.row_id == "8c", f"expected row 8c, got {result.row_id!r}"
-
-    def test_recorded_verdict_unaffected_routes_via_row8(self):
-        """A recorded CHANGES REQUESTED verdict still routes via row 8 — not 8c."""
-        states = self._dead_end_states()
-        states["REVIEW"] = "in_progress"
-        states["_verdicts"] = {
-            "REVIEW": {"verdict": "CHANGES REQUESTED", "recorded_at": "2026-06-14T10:45:00+00:00"}
-        }
-        meta = {
-            "issue_number": 1687,
-            "pr_number": 9999,
-            "latest_review_verdict": "CHANGES REQUESTED",
-            "last_dispatched_skill": SKILL_DO_PR_REVIEW,
-        }
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PATCH
-        assert result.row_id == "8", f"expected row 8, got {result.row_id!r}"
-
-    def test_ordering_8b_before_8c_before_9(self):
-        """Row ordering: 8b must precede 8c, which must precede 9 in DISPATCH_RULES."""
-        from agent.sdlc_router import DISPATCH_RULES
-
-        row_ids = [r.row_id for r in DISPATCH_RULES]
-        idx_8b = row_ids.index("8b")
-        idx_8c = row_ids.index("8c")
-        idx_9 = row_ids.index("9")
-        assert idx_8b < idx_8c, f"8b ({idx_8b}) must come before 8c ({idx_8c})"
-        assert idx_8c < idx_9, f"8c ({idx_8c}) must come before 9 ({idx_9})"
-
-
-class TestNeedsRevisionInvalidatedByRevision:
-    """WS4 (#2049): a plan revision INVALIDATES a NEEDS REVISION verdict.
-
-    The #1760 latch exists to protect the settle-and-build path: a READY TO
-    BUILD (with concerns) verdict stays fresh across its own settle revision
-    so row 4c can route to BUILD. But the latch previously engaged for EVERY
-    verdict kind — including NEEDS REVISION, where the requested revision is
-    exactly what invalidates the verdict. Suppressing staleness there made
-    row 2b step aside and row 3 re-dispatch ``/do-plan`` forever (the
-    #1925/#1968 deadlock, recurring because ``/do-plan`` re-writes
-    ``revision_applied_at`` on every pass, re-arming the suppression).
-
-    Timestamp-only fix: the latch consumes ONLY ``revision_applied_at`` and
-    now engages solely for non-revision-requiring verdicts. No boolean
-    fallback exists on any path.
-    """
-
-    @staticmethod
-    def _needs_revision_state(verdict_at, plan_dispatch_at, revision_applied_at):
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "in_progress",
-            "_verdicts": {"CRITIQUE": {"verdict": "NEEDS REVISION", "recorded_at": verdict_at}},
-            "_sdlc_dispatches": [{"skill": "/do-plan", "at": plan_dispatch_at}],
-        }
-        meta = {
-            "last_dispatched_skill": SKILL_DO_PLAN,
-            "latest_critique_verdict": "NEEDS REVISION",
-            "revision_applied": True,
-            "revision_applied_at": revision_applied_at,
-        }
-        return states, meta
-
-    def test_settled_revision_routes_to_re_critique_first_round(self):
-        """critique(NEEDS REVISION) → /do-plan revision (co-writes
-        revision_applied_at) → next dispatch is /do-plan-critique, never
-        /do-plan again."""
-        states, meta = self._needs_revision_state(
-            "2026-07-13T10:00:00",  # verdict T1
-            "2026-07-13T10:10:00",  # /do-plan revision T2 > T1
-            "2026-07-13T10:20:00",  # revision_applied_at T3 >= T2
-        )
-        assert _critique_verdict_is_stale(states, meta) is True
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-        assert result.skill != SKILL_DO_PLAN
-
-    def test_settled_revision_routes_to_re_critique_second_round(self):
-        """The convergence must hold twice in a row: a SECOND NEEDS REVISION
-        verdict followed by a second settled revision re-routes to
-        /do-plan-critique again (no deadlock on round 2 either)."""
-        states, meta = self._needs_revision_state(
-            "2026-07-13T11:00:00",  # round-2 verdict T4
-            "2026-07-13T11:10:00",  # round-2 /do-plan revision T5 > T4
-            "2026-07-13T11:20:00",  # round-2 revision_applied_at T6 >= T5
-        )
-        # Round-1 history precedes round 2 in the dispatch list.
-        states["_sdlc_dispatches"] = [
-            {"skill": "/do-plan", "at": "2026-07-13T10:10:00"},
-            {"skill": "/do-plan-critique", "at": "2026-07-13T10:30:00"},
-            {"skill": "/do-plan", "at": "2026-07-13T11:10:00"},
-        ]
-        assert _critique_verdict_is_stale(states, meta) is True
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-
-    def test_fresh_needs_revision_still_routes_to_do_plan(self):
-        """Inverse: a NEEDS REVISION verdict with NO revision yet (verdict is
-        the latest event) is NOT stale — row 3 (or G1) still routes to
-        /do-plan for the revision itself."""
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "NEEDS REVISION",
-                    "recorded_at": "2026-07-13T10:00:00",
-                }
-            },
-            "_sdlc_dispatches": [{"skill": "/do-plan", "at": "2026-07-13T09:00:00"}],
-        }
-        meta = {
-            "last_dispatched_skill": SKILL_DO_PLAN_CRITIQUE,
-            "latest_critique_verdict": "NEEDS REVISION",
-            "revision_applied": False,
-        }
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN
-
-    def test_1760_with_concerns_settle_is_now_recritiqued_below_the_bound(self):
-        """#2787 supersedes #1760 on the with-concerns path — the flip.
-
-        #1760 suppressed row 2b here because the loop had no terminating bound,
-        and #2049 then narrowed the latch AWAY from NEEDS REVISION, leaving
-        READY TO BUILD (with concerns) as its live domain. The bound now exists,
-        so below it the concern-closing revision is judged instead of built on.
-
-        The inverse guarantee has NOT been abandoned, it has moved: it is
-        restored permanently once the bound is spent (asserted below), and the
-        no-concerns path is untouched (TestConvergenceLatchRevisionAppliedAt).
-        """
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "READY TO BUILD (with concerns)",
-                    "recorded_at": "2026-07-13T10:00:00",
-                }
-            },
-            "_sdlc_dispatches": [{"skill": "/do-plan", "at": "2026-07-13T10:10:00"}],
-        }
-        meta = {
-            "last_dispatched_skill": SKILL_DO_PLAN,
-            "latest_critique_verdict": "READY TO BUILD (with concerns)",
-            "revision_applied": True,
-            "revision_applied_at": "2026-07-13T10:20:00",
-            "concern_round_count": 1,
-        }
-        assert _critique_verdict_is_stale(states, meta) is True
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-        assert result.row_id == "2b"
-
-    def test_1760_inverse_guarantee_restored_at_the_bound(self):
-        """At the bound the latch engages again and the build proceeds."""
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "READY TO BUILD (with concerns)",
-                    "recorded_at": "2026-07-13T10:00:00",
-                }
-            },
-            "_sdlc_dispatches": [{"skill": "/do-plan", "at": "2026-07-13T10:10:00"}],
-        }
-        meta = {
-            "last_dispatched_skill": SKILL_DO_PLAN,
-            "latest_critique_verdict": "READY TO BUILD (with concerns)",
-            "revision_applied": True,
-            "revision_applied_at": "2026-07-13T10:20:00",
-            "concern_round_count": 3,
-        }
-        assert _critique_verdict_is_stale(states, meta) is False
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_BUILD
-        assert result.row_id == "4c"
-
-    def test_no_boolean_fallback_bare_revision_applied_changes_nothing(self):
-        """Timestamp-only: with revision_applied=True but NO
-        revision_applied_at, the latch is inert on both verdict kinds — the
-        sticky boolean alone never suppresses staleness (no free pass to
-        BUILD, no suppression of re-critique)."""
-        states, meta = self._needs_revision_state(
-            "2026-07-13T10:00:00",
-            "2026-07-13T10:10:00",
-            None,
-        )
-        del meta["revision_applied_at"]
-        assert meta["revision_applied"] is True
-        # Latch inert -> plain timestamp staleness -> stale -> re-critique.
-        assert _critique_verdict_is_stale(states, meta) is True
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-
-
-# ---------------------------------------------------------------------------
-# Row 8b widening — stale REVIEW verdict ownership (#2767b)
-# ---------------------------------------------------------------------------
-
-_VERDICT_AT = "2026-08-13T10:00:00+00:00"
-_PATCH_BEFORE_VERDICT = "2026-08-13T09:00:00+00:00"
-_PATCH_AFTER_VERDICT = "2026-08-13T11:00:00+00:00"
-
-
-def _post_patch_states(
-    *,
-    verdict: str | None = "CHANGES REQUESTED",
-    recorded_at: str | None = _VERDICT_AT,
-    patch_dispatched_at: str | None = _PATCH_AFTER_VERDICT,
-    review_status: str = "completed",
-) -> dict:
-    """A PR-stage state after build → review → patch.
-
-    ``patch_dispatched_at`` controls staleness: a ``/do-patch`` dispatch
-    recorded AFTER ``recorded_at`` makes the verdict stale.
-    """
-    verdicts: dict = {}
-    if verdict is not None:
-        record: dict = {"verdict": verdict}
-        if recorded_at is not None:
-            record["recorded_at"] = recorded_at
-        verdicts["REVIEW"] = record
-
-    dispatches: list = [{"skill": SKILL_DO_BUILD, "at": "2026-08-13T08:00:00+00:00"}]
-    if patch_dispatched_at is not None:
-        dispatches.append({"skill": SKILL_DO_PATCH, "at": patch_dispatched_at})
-
-    return {
-        "ISSUE": "ready",
-        "PLAN": "completed",
-        "CRITIQUE": "completed",
-        "BUILD": "completed",
-        "TEST": "completed",
-        "PATCH": "completed",
-        "REVIEW": review_status,
-        "DOCS": "pending",
-        "MERGE": "pending",
-        "_verdicts": verdicts,
-        "_sdlc_dispatches": dispatches,
-    }
-
-
-class TestRow8bOwnsStaleVerdict:
-    """The spike-1 four-scenario table (#2767b). All four must Dispatch.
-
-    Before the widening, the fourth row (stale verdict + last dispatch
-    ``/do-pr-review``) was owned by nobody and returned
-    ``Blocked('no matching dispatch rule')``.
-    """
-
-    def test_fresh_changes_requested_last_patch_routes_to_patch(self):
-        states = _post_patch_states(patch_dispatched_at=_PATCH_BEFORE_VERDICT)
-        meta = {"pr_number": 4242, "last_dispatched_skill": SKILL_DO_PATCH}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch), f"expected Dispatch, got {result!r}"
-        assert result.skill == SKILL_DO_PATCH
-        assert result.row_id == "8"
-
-    def test_fresh_changes_requested_last_review_routes_to_patch(self):
-        states = _post_patch_states(patch_dispatched_at=_PATCH_BEFORE_VERDICT)
-        meta = {"pr_number": 4242, "last_dispatched_skill": SKILL_DO_PR_REVIEW}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch), f"expected Dispatch, got {result!r}"
-        assert result.skill == SKILL_DO_PATCH
-        assert result.row_id == "8"
-
-    def test_no_verdict_at_all_routes_to_review(self):
-        for last in (SKILL_DO_PATCH, SKILL_DO_PR_REVIEW):
-            states = _post_patch_states(verdict=None, review_status="pending")
-            meta = {"pr_number": 4242, "last_dispatched_skill": last}
-            result = decide_next_dispatch(states, meta)
-            assert isinstance(result, Dispatch), (
-                f"expected Dispatch for last={last}, got {result!r}"
-            )
-            assert result.skill == SKILL_DO_PR_REVIEW
-            assert result.row_id == "7"
-
-    def test_stale_changes_requested_last_patch_routes_to_review(self):
-        states = _post_patch_states()
-        meta = {"pr_number": 4242, "last_dispatched_skill": SKILL_DO_PATCH}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch), f"expected Dispatch, got {result!r}"
-        assert result.skill == SKILL_DO_PR_REVIEW
-        assert result.row_id == "8b"
-
-    def test_stale_changes_requested_last_review_routes_to_review(self):
-        """THE #2767b BUG: previously Blocked('no matching dispatch rule')."""
-        states = _post_patch_states()
-        meta = {"pr_number": 4242, "last_dispatched_skill": SKILL_DO_PR_REVIEW}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch), f"expected Dispatch, got {result!r}"
-        assert result.skill == SKILL_DO_PR_REVIEW
-        assert result.row_id == "8b"
-
-    def test_stale_verdict_owned_regardless_of_last_dispatched_skill(self):
-        """Even an unrelated/absent last dispatch leaves the state owned."""
-        for last in (SKILL_DO_BUILD, "", None):
-            states = _post_patch_states()
-            meta = {"pr_number": 4242, "last_dispatched_skill": last}
-            result = decide_next_dispatch(states, meta)
-            assert isinstance(result, Dispatch), f"expected Dispatch for last={last!r}"
-            assert result.row_id == "8b"
-
-
-class TestRows8And9Unchanged:
-    """The widening must not disturb the fresh-verdict paths."""
-
-    def test_fresh_changes_requested_still_row_8(self):
-        states = _post_patch_states(patch_dispatched_at=None)
-        meta = {"pr_number": 4242, "last_dispatched_skill": SKILL_DO_PR_REVIEW}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PATCH
-        assert result.row_id == "8"
-
-    def test_fresh_approved_review_completed_still_row_9(self):
-        states = _post_patch_states(verdict="APPROVED", patch_dispatched_at=None)
-        meta = {"pr_number": 4242, "last_dispatched_skill": SKILL_DO_PR_REVIEW}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_DOCS
-        assert result.row_id == "9"
-
-
-class TestStaleApprovedIsReReviewedNotAdvanced:
-    """The widening's one genuine ownership TRANSFER, pinned deliberately.
-
-    Row 8b precedes rows 8f/9/10 in ``DISPATCH_RULES``, so a **stale** APPROVED
-    verdict -- one recorded before the latest ``/do-patch`` dispatch -- is now
-    re-reviewed rather than advanced to ``/do-docs`` or the merge fast-path.
-
-    This is not incidental. An APPROVED verdict that predates a patch does not
-    describe the code the patch produced, and advancing on it would merge code
-    no review ever saw. Spike-2's "rows 8 and 9 unchanged" holds for FRESH
-    verdicts only; these cases are the boundary of that claim.
-
-    It converges: a re-review records a verdict newer than the patch dispatch,
-    which is then fresh, and row 9 owns it again. G4 bounds the loop if it
-    somehow does not.
-    """
-
-    def test_stale_approved_review_completed_routes_to_re_review(self):
-        states = _post_patch_states(verdict="APPROVED", review_status="completed")
-        meta = {"pr_number": 4242, "last_dispatched_skill": SKILL_DO_PR_REVIEW}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch), f"expected Dispatch, got {result!r}"
-        assert result.skill == SKILL_DO_PR_REVIEW
-        assert result.row_id == "8b"
-
-    def test_a_fresh_re_review_then_advances_to_docs(self):
-        """Convergence, executable: the same state with a verdict recorded
-        AFTER the patch dispatch is fresh, and row 9 owns it."""
-        states = _post_patch_states(
-            verdict="APPROVED", review_status="completed", patch_dispatched_at=None
-        )
-        meta = {"pr_number": 4242, "last_dispatched_skill": SKILL_DO_PR_REVIEW}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_DOCS
-        assert result.row_id == "9"
-
-
-class TestRow8bDisjointFromEmptyVerdictRows:
-    """Rows 8c/8d/8e require NO recorded verdict; the new staleness disjunct is
-    identically False there because ``recorded_at`` only exists alongside a
-    recorded verdict (``tools.sdlc_verdict.record_verdict`` refuses to write an
-    empty verdict)."""
-
-    def test_row_8c_in_progress_no_verdict_still_wins(self):
-        states = _post_patch_states(verdict=None, review_status="in_progress")
-        states["PATCH"] = "pending"
-        meta = {"pr_number": 4242, "last_dispatched_skill": SKILL_DO_PR_REVIEW}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.row_id == "8c"
-
-    def test_row_8d_crash_after_dispatch_still_wins(self):
-        states = _post_patch_states(verdict=None, review_status="failed")
-        meta = {"pr_number": 4242, "last_dispatched_skill": SKILL_DO_PR_REVIEW}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.row_id == "8d"
-
-    def test_row_8e_completed_no_verdict_still_wins(self):
-        states = _post_patch_states(verdict=None, review_status="completed")
-        meta = {"pr_number": 4242, "last_dispatched_skill": SKILL_DO_BUILD}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.row_id == "8e"
-
-    def test_staleness_disjunct_is_false_on_every_empty_verdict_state(self):
-        """The proof, executable: no recorded verdict → never stale."""
-        for review_status in ("in_progress", "failed", "completed", "pending"):
-            states = _post_patch_states(verdict=None, review_status=review_status)
-            assert _review_verdict_is_stale(states) is False
-
-
-class TestRow8bFailurePaths:
-    def test_malformed_recorded_at_is_not_stale_and_row_8_keeps_control(self):
-        """``_review_verdict_is_stale`` fails safe to False on a garbage
-        timestamp, so a fresh-looking CHANGES REQUESTED verdict still routes to
-        ``/do-patch`` rather than spinning in a re-review loop."""
-        states = _post_patch_states(recorded_at="not-a-timestamp")
-        assert _review_verdict_is_stale(states) is False
-        meta = {"pr_number": 4242, "last_dispatched_skill": SKILL_DO_PR_REVIEW}
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Dispatch)
-        assert result.skill == SKILL_DO_PATCH
-        assert result.row_id == "8"
-
-    def test_predicate_never_raises_on_malformed_verdicts_payload(self):
-        malformed = [
-            {"_verdicts": None},
-            {"_verdicts": "nonsense"},
-            {"_verdicts": []},
-            {"_verdicts": {"REVIEW": None}},
-            {"_verdicts": {"REVIEW": "legacy bare string"}},
-            {"_verdicts": {"REVIEW": {}}},
-            {"_verdicts": {"REVIEW": {"recorded_at": None}}},
-            {"_verdicts": {"REVIEW": {"recorded_at": 12345}}},
-            {"_verdicts": {"REVIEW": ["not", "a", "dict"]}},
-        ]
-        meta = {"pr_number": 4242, "last_dispatched_skill": SKILL_DO_PR_REVIEW}
-        for extra in malformed:
-            states = _post_patch_states()
-            states.update(extra)
-            # Must not raise, and must answer a plain bool.
-            assert _rule_patch_applied_after_review(states, meta, {}) in (True, False), (
-                f"row 8b predicate misbehaved on {extra!r}"
-            )
-
-
-class TestNoRuleBlockIsDistinguishable:
-    """The no-rule fallthrough carries the ``NO_RULE`` sentinel (#2767b) so a
-    caller can tell "the table has a hole" from "a numbered guard fired"
-    without parsing the reason string."""
-
-    def _unowned_state(self) -> tuple[dict, dict]:
-        # PR present, REVIEW completed with an APPROVED verdict, DOCS completed,
-        # MERGE completed — every PR-stage row has already been satisfied.
-        states = {
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "completed",
-            "REVIEW": "completed",
-            "DOCS": "completed",
-            "MERGE": "completed",
-            "_verdicts": {"REVIEW": {"verdict": "APPROVED", "recorded_at": _VERDICT_AT}},
-        }
-        meta = {
-            "pr_number": 4242,
-            "last_dispatched_skill": SKILL_DO_MERGE,
-            # Resolved mergeability, so the separate UNKNOWN-merge-state Blocked
-            # branch stays out of the way and the fallthrough is what fires.
-            "pr_merge_state": "BLOCKED",
-        }
-        return states, meta
-
-    def test_no_rule_block_uses_no_rule_sentinel(self):
-        states, meta = self._unowned_state()
-        result = decide_next_dispatch(states, meta)
-        assert isinstance(result, Blocked), f"expected Blocked, got {result!r}"
-        assert result.reason == "no matching dispatch rule"
-        assert result.guard_id == "NO_RULE"
-
-    def test_no_rule_block_renders_distinguishably_from_a_guard_block(self):
-        states, meta = self._unowned_state()
-        no_rule = decide_next_dispatch(states, meta)
-        guard = decide_next_dispatch(
-            {},
-            {"same_stage_dispatch_count": 99, "last_dispatched_skill": SKILL_DO_PR_REVIEW},
-        )
-        assert isinstance(no_rule, Blocked)
-        assert isinstance(guard, Blocked)
-        assert no_rule.guard_id != guard.guard_id
-        assert no_rule.reason != guard.reason
-        assert guard.guard_id == "G4"
-
-
-# ---------------------------------------------------------------------------
-# #2787: with-concerns re-critique gate.
-#
-# The must-pass gate set from the plan's task 10a. The G5 alive states are the
-# load-bearing ones: guards run to completion BEFORE the dispatch table, and
-# `CRITIQUE_READY_TO_BUILD in verdict_text` matches "READY TO BUILD (WITH
-# CONCERNS)" too, so without G5's step-aside every row below is unreachable in
-# production while still passing its own unit tests.
-# ---------------------------------------------------------------------------
-
-_WC = "READY TO BUILD (with concerns)"
-_PLAN_HASH = "sha256:cafe"
-
-
-def _wc_states(recorded_at: str, plan_dispatch_at: str | None = None) -> dict:
-    """Ledger with a with-concerns CRITIQUE verdict whose hash matches the plan."""
-    states = {
-        "ISSUE": "completed",
-        "PLAN": "completed",
-        "CRITIQUE": "completed",
-        "BUILD": "pending",
-        "_verdicts": {
-            "CRITIQUE": {
-                "verdict": "READY TO BUILD (WITH CONCERNS)",
-                "recorded_at": recorded_at,
-                "artifact_hash": _PLAN_HASH,
-            }
-        },
-    }
-    if plan_dispatch_at:
-        states["_sdlc_dispatches"] = [
-            {"skill": SKILL_DO_PLAN, "at": plan_dispatch_at, "stage": "PLAN"}
-        ]
-    return states
-
-
-def _wc_meta(revision_applied_at: str | None, count: int = 0, **extra) -> dict:
-    meta = {
-        "latest_critique_verdict": _WC,
-        "revision_applied_at": revision_applied_at,
-        "concern_round_count": count,
-        # Sticky and deliberately WRONG for the routing under test: every row
-        # keyed on it before #2787 sent round 2+ straight to /do-build.
-        "revision_applied": True,
-    }
-    meta.update(extra)
-    return meta
-
-
-class TestG5AliveOnWithConcerns:
-    """G5 must never serve a with-concerns verdict, in any revision state."""
-
-    def _ctx(self):
-        return {"current_plan_hash": _PLAN_HASH}
-
-    def test_a_state_s1_unlocked_routes_to_row_4b(self):
-        """S1: verdict is newest, no revision since, lock cleared -> row 4b."""
-        states = _wc_states(_iso("2026-08-17T02:00:00"))
-        meta = _wc_meta(_iso("2026-08-17T01:00:00"))
-        result = decide_next_dispatch(states, meta, self._ctx())
-        assert result.row_id != "G5"
-        assert result.skill == SKILL_DO_PLAN
-        assert result.row_id == "4b"
-
-    def test_a2_state_s1_with_step_5_6_lock_routes_to_do_plan(self):
-        """S1 as Step 5.6 actually leaves it: the lock is set.
-
-        G7 gate 4 legitimately owns this turn and dispatches /do-plan with
-        row_id="G7" before the dispatch table is consulted. Assert the SKILL,
-        not the row — but still pin that G5 did not ship a build.
-        """
-        states = _wc_states(_iso("2026-08-17T02:00:00"))
-        meta = _wc_meta(
-            _iso("2026-08-17T01:00:00"),
-            plan_revising=True,
-            last_dispatched_skill=SKILL_DO_PLAN_CRITIQUE,
-        )
-        result = decide_next_dispatch(states, meta, self._ctx())
-        assert result.row_id != "G5"
-        assert result.skill == SKILL_DO_PLAN
-        assert result.row_id in {"4b", "G7"}
-
-    def test_a3_plan_dispatched_but_no_revision_landed_stays_on_row_4b(self):
-        """Row 4b fired, /do-plan crashed before writing revision_applied_at.
-
-        The dispatch record's `at` postdates the verdict, so a fallthrough to
-        `verdict_dt < plan_dt` would call the verdict stale and let row 2b
-        re-critique a plan nobody revised. It must resolve to row 4b instead.
-        """
-        states = _wc_states(
-            _iso("2026-08-17T02:00:00"),
-            plan_dispatch_at=_iso("2026-08-17T02:30:00"),
-        )
-        meta = _wc_meta(_iso("2026-08-17T01:00:00"))
-        result = decide_next_dispatch(states, meta, self._ctx())
-        assert result.row_id != "G5"
-        assert result.row_id != "2b"
-        assert result.skill == SKILL_DO_PLAN
-        assert result.row_id == "4b"
-
-    def test_b_state_s2_below_bound_routes_to_row_2b(self):
-        """S2 below the bound: the revision landed and must be re-critiqued."""
-        states = _wc_states(
-            _iso("2026-08-17T02:00:00"),
-            plan_dispatch_at=_iso("2026-08-17T02:30:00"),
-        )
-        meta = _wc_meta(_iso("2026-08-17T03:00:00"), count=1)
-        result = decide_next_dispatch(states, meta, self._ctx())
-        assert result.row_id != "G5"
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-        assert result.row_id == "2b"
-
-    def test_c_state_s2_at_bound_routes_to_row_4c(self):
-        """S2 with the bound spent: build, with the acceptance in the reason."""
-        states = _wc_states(
-            _iso("2026-08-17T02:00:00"),
-            plan_dispatch_at=_iso("2026-08-17T02:30:00"),
-        )
-        meta = _wc_meta(_iso("2026-08-17T03:00:00"), count=3)
-        result = decide_next_dispatch(states, meta, self._ctx())
-        assert result.row_id != "G5"
-        assert result.skill == SKILL_DO_BUILD
-        assert result.row_id == "4c"
-        assert "bound" in result.reason.lower()
-        assert "accepted" in result.reason.lower()
-
-    def test_g5_still_fires_on_a_no_concerns_cache_hit(self):
-        """The step-aside must not disarm G5 for clean verdicts."""
-        states = {
-            "ISSUE": "completed",
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "pending",
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "READY TO BUILD",
-                    "recorded_at": _iso("2026-08-17T02:00:00"),
-                    "artifact_hash": _PLAN_HASH,
-                }
-            },
-        }
-        meta = {"latest_critique_verdict": "READY TO BUILD"}
-        result = decide_next_dispatch(states, meta, self._ctx())
-        assert result.skill == SKILL_DO_BUILD
-        assert result.row_id == "G5"
-
-
-class TestConcernBoundScoping:
-    """The bound counts with-concerns rounds only, and terminates the loop."""
-
-    def test_a_lane_with_a_long_needs_revision_history_still_gets_re_critique(self):
-        """A NEEDS REVISION history must not arrive at the loop with the bound spent.
-
-        The counter is written only on a WITH CONCERNS verdict, so `count` is 1
-        on the first with-concerns round no matter how many NEEDS REVISION
-        rounds preceded it. Here the ledger carries a dispatch history of five
-        prior /do-plan-critique turns; routing must still be row 2b, not the
-        bound-exhausted row 4c. (That the *counter* ignores NEEDS REVISION is
-        proven writer-side in
-        tests/unit/test_sdlc_verdict.py::TestConcernRoundCounter.)
-        """
-        states = _wc_states(
-            _iso("2026-08-17T02:00:00"),
-            plan_dispatch_at=_iso("2026-08-17T02:30:00"),
-        )
-        states["_sdlc_dispatches"] = [
-            {"skill": SKILL_DO_PLAN_CRITIQUE, "at": _iso("2026-08-17T00:00:00")},
-            {"skill": SKILL_DO_PLAN, "at": _iso("2026-08-17T00:10:00")},
-            {"skill": SKILL_DO_PLAN_CRITIQUE, "at": _iso("2026-08-17T00:20:00")},
-            {"skill": SKILL_DO_PLAN, "at": _iso("2026-08-17T00:30:00")},
-            {"skill": SKILL_DO_PLAN_CRITIQUE, "at": _iso("2026-08-17T00:40:00")},
-            {"skill": SKILL_DO_PLAN, "at": _iso("2026-08-17T02:30:00")},
-        ]
-        meta = _wc_meta(_iso("2026-08-17T03:00:00"), count=1)
-        result = decide_next_dispatch(states, meta, {"current_plan_hash": _PLAN_HASH})
-        assert result.skill == SKILL_DO_PLAN_CRITIQUE
-        assert result.row_id == "2b"
-
-    def test_mixed_timezone_awareness_falls_safe_to_a_revision_pass(self):
-        """A naive frontmatter timestamp vs a tz-aware recorded_at must not build.
-
-        `recorded_at` is always tz-aware (`datetime.now(UTC).isoformat()`), while
-        `revision_applied_at` is parsed out of hand-editable plan frontmatter and
-        may be naive. Comparing the two raises TypeError; the predicate swallows
-        it to False, so the lane routes to a revision pass rather than to a build.
-        """
-        states = _wc_states("2026-08-17T02:00:00+00:00")
-        meta = _wc_meta("2026-08-17T03:00:00", count=3)
-        assert _concern_revision_is_unjudged(states, meta) is False
-        result = decide_next_dispatch(states, meta, {"current_plan_hash": _PLAN_HASH})
-        assert result.skill == SKILL_DO_PLAN
-        assert result.row_id == "4b"
-
-    def test_no_concerns_verdict_with_settled_revision_still_routes_to_4a(self):
-        """The with-concerns branch must not leak into the clean path."""
-        states = {
-            "ISSUE": "completed",
-            "PLAN": "completed",
-            "CRITIQUE": "completed",
-            "BUILD": "pending",
-            "_verdicts": {
-                "CRITIQUE": {
-                    "verdict": "READY TO BUILD",
-                    "recorded_at": _iso("2026-08-17T02:00:00"),
-                }
-            },
-        }
-        meta = {
-            "latest_critique_verdict": "READY TO BUILD",
-            "revision_applied_at": _iso("2026-08-17T03:00:00"),
-        }
-        result = decide_next_dispatch(states, meta)
-        assert result.skill == SKILL_DO_BUILD
-        assert result.row_id == "4a"
-
-    def test_with_concerns_no_plan_dispatch_falls_safe_to_row_4b(self):
-        """Bound spent but no /do-plan dispatch recorded and no revision landed.
-
-        Pins that the with-concerns decision is control-flow independent of
-        `_sdlc_dispatches`: the branch sits ahead of the `latest_plan_at` early
-        return. Fail-safe direction is a revision pass, never a build.
-        """
-        states = _wc_states(_iso("2026-08-17T02:00:00"))
-        meta = _wc_meta(None, count=3)
-        result = decide_next_dispatch(states, meta, {"current_plan_hash": _PLAN_HASH})
-        assert result.skill == SKILL_DO_PLAN
-        assert result.row_id == "4b"
-
-    def test_kill_switch_restores_pre_2787_routing(self, monkeypatch):
-        """MAX_CONCERN_RECRITIQUE_ROUNDS=0 keeps the latch permanently engaged."""
-        import agent.sdlc_router as router
-
-        monkeypatch.setattr(router, "MAX_CONCERN_RECRITIQUE_ROUNDS", 0)
-        states = _wc_states(
-            _iso("2026-08-17T02:00:00"),
-            plan_dispatch_at=_iso("2026-08-17T02:30:00"),
-        )
-        meta = _wc_meta(_iso("2026-08-17T03:00:00"), count=0)
-        assert _critique_verdict_is_stale(states, meta) is False
-        result = decide_next_dispatch(states, meta, {"current_plan_hash": _PLAN_HASH})
-        assert result.skill == SKILL_DO_BUILD
-        assert result.row_id == "4c"
-
-
-class TestDispatchRuleOrderingIsLoadBearing:
-    def test_row_2b_precedes_rows_4b_and_4c(self):
-        """The design depends on first-match ordering; pin it with a test."""
-        ids = [r.row_id for r in DISPATCH_RULES]
-        assert ids.index("2b") < ids.index("4b")
-        assert ids.index("2b") < ids.index("4c")
-
-
-class TestForeverWithConcernsTerminates:
-    """#2787 gate item 5: the loop terminates, and says so out loud.
-
-    ``compute_same_stage_count`` breaks its streak on every skill change and
-    this loop alternates ``/do-plan`` and ``/do-plan-critique``, so G4 can
-    never fire on it. ``MAX_CONCERN_RECRITIQUE_ROUNDS`` is the ONLY
-    terminator. This simulation is the proof that it actually terminates: it
-    drives the real router turn by turn against a plan whose critique returns
-    ``READY TO BUILD (with concerns)`` forever.
-    """
-
-    def _simulate(self, max_turns: int = 40) -> list:
-        """Drive decide_next_dispatch, applying each dispatch's real effect.
-
-        - ``/do-plan-critique`` records a fresh with-concerns verdict:
-          ``recorded_at`` advances and ``record_verdict`` bumps the counter.
-        - ``/do-plan`` settles a revision: ``revision_applied_at`` advances.
-        - ``/do-build`` is terminal.
-        """
-        clock = [0]
-
-        def tick() -> str:
-            clock[0] += 1
-            return _iso(f"2026-08-17T{clock[0]:02d}:00:00")
-
-        states = _wc_states(tick())
-        meta = _wc_meta(None, count=1)  # the first verdict already counted
-        trail = []
-        for _ in range(max_turns):
-            result = decide_next_dispatch(states, meta, {"current_plan_hash": _PLAN_HASH})
-            assert isinstance(result, Dispatch), f"loop escalated instead of terminating: {result}"
-            trail.append(result)
-            if result.skill == SKILL_DO_BUILD:
-                return trail
-            if result.skill == SKILL_DO_PLAN:
-                # /do-plan Phase 4 step 2a co-writes both fields.
-                meta["revision_applied"] = True
-                meta["revision_applied_at"] = tick()
-            elif result.skill == SKILL_DO_PLAN_CRITIQUE:
-                # A fresh with-concerns verdict: new recorded_at, counter +1.
-                states["_verdicts"]["CRITIQUE"]["recorded_at"] = tick()
-                meta["concern_round_count"] += 1
-            else:  # pragma: no cover - defensive
-                raise AssertionError(f"unexpected skill in the loop: {result.skill}")
-            # The router's own dispatch record, as record_dispatch would write it.
-            states.setdefault("_sdlc_dispatches", []).append(
-                {"skill": result.skill, "at": _iso(f"2026-08-17T{clock[0]:02d}:00:00")}
-            )
-            meta["last_dispatched_skill"] = result.skill
-        raise AssertionError(f"never terminated in {max_turns} turns: {[d.row_id for d in trail]}")
-
-    def test_terminates_at_the_bound_with_a_build(self):
-        trail = self._simulate()
-        assert trail[-1].skill == SKILL_DO_BUILD
-        assert trail[-1].row_id == "4c"
-        # It must re-critique at least once before giving up, or the bound is
-        # not doing the job the feature exists to do.
-        assert any(d.row_id == "2b" for d in trail), [d.row_id for d in trail]
-
-    def test_consumes_exactly_the_bounded_number_of_recritiques(self):
-        """Round 1's verdict is already counted, so the loop buys the rest."""
-        trail = self._simulate()
-        recritiques = [d for d in trail if d.skill == SKILL_DO_PLAN_CRITIQUE]
-        assert len(recritiques) == MAX_CONCERN_RECRITIQUE_ROUNDS - 1, [d.row_id for d in trail]
-
-    def test_row_4c_reason_records_the_accepted_residual_concerns(self):
-        """Gate item 5: a silent build at the cap is the failure mode one level up.
-
-        The reason string is the accountability record the supervisor sees, so
-        assert the string itself — the bound, the count, and the fact that
-        residual concerns were accepted WITHOUT review.
-        """
-        reason = self._simulate()[-1].reason
-        lowered = reason.lower()
-        assert "bound" in lowered, reason
-        assert str(MAX_CONCERN_RECRITIQUE_ROUNDS) in reason, reason
-        assert "with-concerns rounds" in lowered, reason
-        assert "residual concerns accepted unreviewed" in lowered, reason
-
-    def test_g4_cannot_see_this_loop(self):
-        """The bound is the only terminator — pinned so nobody deletes it.
-
-        A future reader will assume G4's oscillation cap backstops this. It
-        does not: ``compute_same_stage_count`` counts CONSECUTIVE same-skill
-        dispatches and breaks the streak on any skill change.
-        """
-        from agent.sdlc_router import compute_same_stage_count
-
-        # An UNCHANGED snapshot throughout: the only thing breaking the streak
-        # is the alternating skill, which is exactly the claim under test.
-        snapshot = {"PLAN": "completed", "CRITIQUE": "completed", "BUILD": "pending"}
-        history = [
-            {
-                "skill": SKILL_DO_PLAN if i % 2 == 0 else SKILL_DO_PLAN_CRITIQUE,
-                "at": _iso("2026-08-17T01:00:00"),
-                "stage_snapshot": snapshot,
-            }
-            for i in range(12)
-        ]
-        count, skill = compute_same_stage_count({"_sdlc_dispatches": history}, snapshot)
-        # +1 for the "about to dispatch" turn; the recorded streak itself is 1.
-        assert count <= 2, (count, skill)
-        assert count < MAX_SAME_STAGE_DISPATCHES, (
-            f"G4 would fire on the alternation at {count}; the bound is supposed "
-            "to be this loop's only terminator"
-        )
